Refactor recursion proving
* Move non-zkvm-specific code from risc0_zkvm::host::recursion::prove to risc0_recursion::prove
* risc0_recursion is po2 agnostic
