+++ risc0/recursion/Cargo.toml
@@ -1,0 +1,63 @@
+[package]
+name = "risc0-recursion"
+description = "RISC Zero recursion interface"
+version = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+homepage = { workspace = true }
+repository = { workspace = true }
+
+[dependencies]
+anyhow = { version = "1.0", default-features = false }
+bytemuck = { version = "1.13", features = ["extern_crate_alloc"] }
+cfg-if = "1.0"
+hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
+risc0-circuit-recursion = { workspace = true }
+risc0-core = { workspace = true }
+risc0-zkp = { workspace = true }
+serde = { version = "1.0", default-features = false, features = [
+  "alloc",
+  "derive",
+] }
+tracing = { version = "0.1", default-features = false, features = [
+  "attributes",
+] }
+
+[target.'cfg(not(target_os = "zkvm"))'.dependencies]
+bytemuck = "1.12"
+cust = { version = "0.3", optional = true }
+lazy-regex = { version = "3.1", optional = true }
+metal = { version = "0.28", optional = true }
+nvtx = { version = "1.3", optional = true }
+rand = { version = "0.8", optional = true }
+rayon = { version = "1.5", optional = true }
+risc0-circuit-recursion-sys = { workspace = true, optional = true }
+risc0-sys = { workspace = true, optional = true }
+sha2 = { version = "0.10", default-features = false }
+zip = { version = "2.0", default-features = false, optional = true, features = [
+  "deflate",
+] }
+
+[dev-dependencies]
+test-log = { version = "0.2", default-features = false, features = ["trace"] }
+tracing-forest = "0.1"
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+
+[features]
+default = []
+prove = [
+  "dep:lazy-regex",
+  "dep:nvtx",
+  "dep:rand",
+  "dep:rayon",
+  "risc0-circuit-recursion/prove",
+  "risc0-zkp/prove",
+  "std",
+]
+std = [
+  "anyhow/std",
+  "hex/std",
+  "risc0-zkp/std",
+  "serde/std",
+  "sha2/std",
+]
