+++ risc0/zkvm/src/host/recursion/prove/mod.rs
@@ -12,62 +12,41 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod exec;
-mod plonk;
-pub mod preflight;
-mod program;
 pub mod zkr;
 
-use std::{collections::VecDeque, fmt::Debug, mem::take, rc::Rc};
+use std::{collections::VecDeque, fmt::Debug};
 
-use anyhow::{anyhow, bail, ensure, Context, Result};
-use rand::thread_rng;
-use risc0_circuit_recursion::{
-    control_id::BN254_IDENTITY_CONTROL_ID, cpu::CpuCircuitHal, CircuitImpl, CIRCUIT,
-    REGISTER_GROUP_ACCUM, REGISTER_GROUP_CTRL, REGISTER_GROUP_DATA,
-};
+use anyhow::{anyhow, ensure, Context, Result};
+use risc0_circuit_recursion::{control_id::BN254_IDENTITY_CONTROL_ID, CircuitImpl};
 use risc0_circuit_rv32im::control_id::POSEIDON2_CONTROL_IDS;
+use risc0_recursion::prove::{DigestKind, RecursionReceipt};
 use risc0_zkp::{
-    adapter::{CircuitInfo, CircuitStepContext, TapsProvider, PROOF_SYSTEM_INFO},
-    core::{
-        digest::Digest,
-        hash::{hash_suite_from_name, poseidon2::Poseidon2HashSuite},
-    },
-    field::{
-        baby_bear::{BabyBear, BabyBearElem, BabyBearExtElem},
-        Elem,
-    },
-    hal::{cpu::CpuHal, CircuitHal, Hal},
-    prove::adapter::ProveAdapter,
+    adapter::{CircuitInfo, PROOF_SYSTEM_INFO},
+    core::{digest::Digest, hash::hash_suite_from_name},
+    field::baby_bear::{BabyBear, BabyBearElem, BabyBearExtElem},
+    hal::{CircuitHal, Hal},
     verify::ReadIOP,
-    MIN_CYCLES_PO2, ZK_CYCLES,
+    MIN_CYCLES_PO2,
 };
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 
-pub use self::program::Program;
 use crate::{
-    receipt::{
-        merkle::{MerkleGroup, MerkleProof},
-        SegmentReceipt, SuccinctReceipt, SuccinctReceiptVerifierParameters,
-    },
+    receipt::{SegmentReceipt, SuccinctReceipt, SuccinctReceiptVerifierParameters},
     receipt_claim::{Assumption, MaybePruned, Merge},
     sha::Digestible,
-    HalPair, ProverOpts, ReceiptClaim,
+    ProverOpts, ReceiptClaim,
 };
 
-// TODO: Automatically generate these constants from the circuit somehow without
-// messing up bootstrap dependencies.
+// Re-export risc0_recursion items that were previously here as to not
+// break dependencies.
+#[allow(unused_imports)]
+use risc0_recursion::{
+    merkle::{MerkleGroup, MerkleProof},
+    prove::Program,
+};
+
 /// Number of rows to use for the recursion circuit witness as a power of 2.
 pub const RECURSION_PO2: usize = 18;
-/// Size of the code group in the taps of the recursion circuit.
-const RECURSION_CODE_SIZE: usize = 23;
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct RecursionReceipt {
-    pub control_id: Digest,
-    pub seal: Vec<u32>,
-    pub output: Vec<u32>,
-}
 
 /// Run the lift program to transform an rv32im segment receipt into a recursion receipt.
 ///
@@ -80,7 +59,7 @@ pub fn lift(segment_receipt: &SegmentReceipt) -> Result<SuccinctReceipt<ReceiptC
     let opts = ProverOpts::succinct();
     let mut prover = Prover::new_lift(segment_receipt, opts.clone())?;
 
-    let receipt = prover.run()?;
+    let receipt = prover.prover.run()?;
     let mut out_stream = VecDeque::<u32>::new();
     out_stream.extend(receipt.output.iter());
     let claim_decoded = ReceiptClaim::decode(&mut out_stream)?;
@@ -88,11 +67,11 @@ pub fn lift(segment_receipt: &SegmentReceipt) -> Result<SuccinctReceipt<ReceiptC
 
     // Include an inclusion proof for control_id to allow verification against a root.
     let control_inclusion_proof = MerkleGroup::new(opts.control_ids.clone())?
-        .get_proof(&receipt.control_id, opts.hash_suite()?.hashfn.as_ref())?;
+        .get_proof(&prover.control_id, opts.hash_suite()?.hashfn.as_ref())?;
     Ok(SuccinctReceipt {
         seal: receipt.seal,
         hashfn: opts.hashfn,
-        control_id: receipt.control_id,
+        control_id: prover.control_id,
         control_inclusion_proof,
         claim: claim_decoded.merge(&segment_receipt.claim)?.into(),
         verifier_parameters: SuccinctReceiptVerifierParameters::default().digest(),
@@ -112,7 +91,7 @@ pub fn join(
 
     let opts = ProverOpts::succinct();
     let mut prover = Prover::new_join(a, b, opts.clone())?;
-    let receipt = prover.run()?;
+    let receipt = prover.prover.run()?;
     let mut out_stream = VecDeque::<u32>::new();
     out_stream.extend(receipt.output.iter());
 
@@ -130,11 +109,11 @@ pub fn join(
 
     // Include an inclusion proof for control_id to allow verification against a root.
     let control_inclusion_proof = MerkleGroup::new(opts.control_ids.clone())?
-        .get_proof(&receipt.control_id, opts.hash_suite()?.hashfn.as_ref())?;
+        .get_proof(&prover.control_id, opts.hash_suite()?.hashfn.as_ref())?;
     Ok(SuccinctReceipt {
         seal: receipt.seal,
         hashfn: opts.hashfn,
-        control_id: receipt.control_id,
+        control_id: prover.control_id,
         control_inclusion_proof,
         claim: claim_decoded.merge(&ab_claim)?.into(),
         verifier_parameters: SuccinctReceiptVerifierParameters::default().digest(),
@@ -192,7 +171,7 @@ where
 
     let opts = ProverOpts::succinct();
     let mut prover = Prover::new_resolve(conditional, assumption, opts.clone())?;
-    let receipt = prover.run()?;
+    let receipt = prover.prover.run()?;
     let mut out_stream = VecDeque::<u32>::new();
     out_stream.extend(receipt.output.iter());
 
@@ -201,11 +180,11 @@ where
 
     // Include an inclusion proof for control_id to allow verification against a root.
     let control_inclusion_proof = MerkleGroup::new(opts.control_ids.clone())?
-        .get_proof(&receipt.control_id, opts.hash_suite()?.hashfn.as_ref())?;
+        .get_proof(&prover.control_id, opts.hash_suite()?.hashfn.as_ref())?;
     Ok(SuccinctReceipt {
         seal: receipt.seal,
         hashfn: opts.hashfn,
-        control_id: receipt.control_id,
+        control_id: prover.control_id,
         control_inclusion_proof,
         claim: claim_decoded.merge(&resolved_claim)?.into(),
         verifier_parameters: SuccinctReceiptVerifierParameters::default().digest(),
@@ -223,7 +202,7 @@ pub fn identity_p254(a: &SuccinctReceipt<ReceiptClaim>) -> Result<SuccinctReceip
         .with_control_ids(vec![BN254_IDENTITY_CONTROL_ID]);
 
     let mut prover = Prover::new_identity(a, opts.clone())?;
-    let receipt = prover.run()?;
+    let receipt = prover.prover.run()?;
     let mut out_stream = VecDeque::<u32>::new();
     out_stream.extend(receipt.output.iter());
     let claim = MaybePruned::Value(ReceiptClaim::decode(&mut out_stream)?).merge(&a.claim)?;
@@ -231,8 +210,8 @@ pub fn identity_p254(a: &SuccinctReceipt<ReceiptClaim>) -> Result<SuccinctReceip
     // Include an inclusion proof for control_id to allow verification against a root.
     let hashfn = opts.hash_suite()?.hashfn;
     let control_inclusion_proof = MerkleGroup::new(opts.control_ids.clone())?
-        .get_proof(&receipt.control_id, hashfn.as_ref())?;
-    let control_root = control_inclusion_proof.root(&receipt.control_id, hashfn.as_ref());
+        .get_proof(&prover.control_id, hashfn.as_ref())?;
+    let control_root = control_inclusion_proof.root(&prover.control_id, hashfn.as_ref());
     let params = SuccinctReceiptVerifierParameters {
         control_root,
         inner_control_root: Some(a.control_root()?),
@@ -242,7 +221,7 @@ pub fn identity_p254(a: &SuccinctReceipt<ReceiptClaim>) -> Result<SuccinctReceip
     Ok(SuccinctReceipt {
         seal: receipt.seal,
         hashfn: opts.hashfn,
-        control_id: receipt.control_id,
+        control_id: prover.control_id,
         control_inclusion_proof,
         claim,
         verifier_parameters: params.digest(),
@@ -262,7 +241,7 @@ pub fn test_recursion_circuit(
     let opts = ProverOpts::succinct().with_control_ids(vec![control_id]);
 
     let mut prover = Prover::new_test_recursion_circuit([digest1, digest2], opts.clone())?;
-    let receipt = prover.run()?;
+    let receipt = prover.prover.run()?;
 
     // Read the claim digest from the second of the global output slots.
     const DIGEST_SHORTS: usize = crate::sha::DIGEST_WORDS * 2;
@@ -278,8 +257,8 @@ pub fn test_recursion_circuit(
     // Include an inclusion proof for control_id to allow verification against a root.
     let hashfn = opts.hash_suite()?.hashfn;
     let control_inclusion_proof = MerkleGroup::new(opts.control_ids.clone())?
-        .get_proof(&receipt.control_id, hashfn.as_ref())?;
-    let control_root = control_inclusion_proof.root(&receipt.control_id, hashfn.as_ref());
+        .get_proof(&prover.control_id, hashfn.as_ref())?;
+    let control_root = control_inclusion_proof.root(&prover.control_id, hashfn.as_ref());
     let params = SuccinctReceiptVerifierParameters {
         control_root,
         inner_control_root: Some(digest1.to_owned()),
@@ -289,171 +268,30 @@ pub fn test_recursion_circuit(
     Ok(SuccinctReceipt {
         seal: receipt.seal,
         hashfn: opts.hashfn,
-        control_id: receipt.control_id,
+        control_id: prover.control_id,
         control_inclusion_proof,
         claim: MaybePruned::Pruned(claim_digest),
         verifier_parameters: params.digest(),
     })
 }
 
-/// Prover for the recursion circuit.
+/// Prover for zkVM use of the recursion circuit.
 pub struct Prover {
-    program: Program,
+    prover: risc0_recursion::prove::Prover,
     control_id: Digest,
-    opts: ProverOpts,
-    input: VecDeque<u32>,
-    split_points: Vec<usize>,
-    output: Vec<u32>,
-}
-
-#[cfg(feature = "cuda")]
-mod cuda {
-    pub use risc0_circuit_recursion::cuda::{CudaCircuitHalPoseidon2, CudaCircuitHalSha256};
-    pub use risc0_zkp::hal::cuda::{CudaHalPoseidon2, CudaHalSha256};
-
-    use super::{HalPair, Rc};
-
-    pub fn sha256_hal_pair() -> HalPair<CudaHalSha256, CudaCircuitHalSha256> {
-        let hal = Rc::new(CudaHalSha256::new());
-        let circuit_hal = Rc::new(CudaCircuitHalSha256::new(hal.clone()));
-        HalPair { hal, circuit_hal }
-    }
-
-    pub fn poseidon2_hal_pair() -> HalPair<CudaHalPoseidon2, CudaCircuitHalPoseidon2> {
-        let hal = Rc::new(CudaHalPoseidon2::new());
-        let circuit_hal = Rc::new(CudaCircuitHalPoseidon2::new(hal.clone()));
-        HalPair { hal, circuit_hal }
-    }
-}
-
-#[cfg(feature = "metal")]
-mod metal {
-    pub use risc0_circuit_recursion::metal::MetalCircuitHal;
-    pub use risc0_zkp::hal::metal::{
-        MetalHalPoseidon2, MetalHalSha256, MetalHashPoseidon2, MetalHashSha256,
-    };
-
-    use super::{HalPair, Rc};
-
-    pub fn sha256_hal_pair() -> HalPair<MetalHalSha256, MetalCircuitHal<MetalHashSha256>> {
-        let hal = Rc::new(MetalHalSha256::new());
-        let circuit_hal = Rc::new(MetalCircuitHal::<MetalHashSha256>::new(hal.clone()));
-        HalPair { hal, circuit_hal }
-    }
-
-    pub fn poseidon2_hal_pair() -> HalPair<MetalHalPoseidon2, MetalCircuitHal<MetalHashPoseidon2>> {
-        let hal = Rc::new(MetalHalPoseidon2::new());
-        let circuit_hal = Rc::new(MetalCircuitHal::<MetalHashPoseidon2>::new(hal.clone()));
-        HalPair { hal, circuit_hal }
-    }
-}
-
-mod cpu {
-    use risc0_zkp::core::hash::{poseidon_254::Poseidon254HashSuite, sha::Sha256HashSuite};
-
-    use super::{
-        BabyBear, CircuitImpl, CpuCircuitHal, CpuHal, HalPair, Poseidon2HashSuite, Rc, CIRCUIT,
-    };
-
-    #[allow(dead_code)]
-    pub fn sha256_hal_pair() -> HalPair<CpuHal<BabyBear>, CpuCircuitHal<'static, CircuitImpl>> {
-        let hal = Rc::new(CpuHal::new(Sha256HashSuite::new_suite()));
-        let circuit_hal = Rc::new(CpuCircuitHal::new(&CIRCUIT));
-        HalPair { hal, circuit_hal }
-    }
-
-    #[allow(dead_code)]
-    pub fn poseidon2_hal_pair() -> HalPair<CpuHal<BabyBear>, CpuCircuitHal<'static, CircuitImpl>> {
-        let hal = Rc::new(CpuHal::new(Poseidon2HashSuite::new_suite()));
-        let circuit_hal = Rc::new(CpuCircuitHal::new(&CIRCUIT));
-        HalPair { hal, circuit_hal }
-    }
-
-    #[allow(dead_code)]
-    pub fn poseidon254_hal_pair() -> HalPair<CpuHal<BabyBear>, CpuCircuitHal<'static, CircuitImpl>>
-    {
-        let hal = Rc::new(CpuHal::new(Poseidon254HashSuite::new_suite()));
-        let circuit_hal = Rc::new(CpuCircuitHal::new(&CIRCUIT));
-        HalPair { hal, circuit_hal }
-    }
-}
-
-cfg_if::cfg_if! {
-    if #[cfg(feature = "cuda")] {
-        /// TODO
-        #[allow(dead_code)]
-        pub fn sha256_hal_pair() -> HalPair<cuda::CudaHalSha256, cuda::CudaCircuitHalSha256> {
-            cuda::sha256_hal_pair()
-        }
-
-        /// TODO
-        #[allow(dead_code)]
-        pub fn poseidon2_hal_pair() -> HalPair<cuda::CudaHalPoseidon2, cuda::CudaCircuitHalPoseidon2> {
-            cuda::poseidon2_hal_pair()
-        }
-
-        /// TODO
-        #[allow(dead_code)]
-        pub fn poseidon254_hal_pair() -> HalPair<CpuHal<BabyBear>, CpuCircuitHal<'static, CircuitImpl>> {
-            cpu::poseidon254_hal_pair()
-        }
-    } else if #[cfg(feature = "metal")] {
-        /// TODO
-        #[allow(dead_code)]
-        pub fn sha256_hal_pair() -> HalPair<metal::MetalHalSha256, metal::MetalCircuitHal<metal::MetalHashSha256>> {
-            metal::sha256_hal_pair()
-        }
-
-        /// TODO
-        #[allow(dead_code)]
-        pub fn poseidon2_hal_pair() -> HalPair<metal::MetalHalPoseidon2, metal::MetalCircuitHal<metal::MetalHashPoseidon2>> {
-            metal::poseidon2_hal_pair()
-        }
-
-        /// TODO
-        #[allow(dead_code)]
-        pub fn poseidon254_hal_pair() -> HalPair<CpuHal<BabyBear>, CpuCircuitHal<'static, CircuitImpl>> {
-            cpu::poseidon254_hal_pair()
-        }
-    } else {
-        /// TODO
-        #[allow(dead_code)]
-        pub fn sha256_hal_pair() -> HalPair<CpuHal<BabyBear>, CpuCircuitHal<'static, CircuitImpl>> {
-            cpu::sha256_hal_pair()
-        }
-
-        /// TODO
-        #[allow(dead_code)]
-        pub fn poseidon2_hal_pair() -> HalPair<CpuHal<BabyBear>, CpuCircuitHal<'static, CircuitImpl>> {
-            cpu::poseidon2_hal_pair()
-        }
-
-        /// TODO
-        #[allow(dead_code)]
-        pub fn poseidon254_hal_pair() -> HalPair<CpuHal<BabyBear>, CpuCircuitHal<'static, CircuitImpl>> {
-            cpu::poseidon254_hal_pair()
-        }
-    }
-}
-
-/// Kinds of digests recognized by the recursion program language.
-// NOTE: Default is additionally a recognized type in the recursion program language. It's not
-// yet supported here because some of the code in this module assumes Poseidon2 is Default.
-enum DigestKind {
-    Poseidon2,
-    Sha256,
 }
 
 impl Prover {
     fn new(program: Program, control_id: Digest, opts: ProverOpts) -> Self {
         Self {
-            program,
+            prover: risc0_recursion::prove::Prover::new(program, &opts.hashfn),
             control_id,
-            opts,
-            input: VecDeque::new(),
-            split_points: Vec::new(),
-            output: Vec::new(),
         }
+    }
+
+    /// Returns the control id of the recursion VM program being proven.
+    pub fn control_id(&self) -> &Digest {
+        &self.control_id
     }
 
     /// Initialize a recursion prover with the test recursion program. This program is used in
@@ -659,25 +497,12 @@ impl Prover {
     }
 
     fn add_input(&mut self, input: &[u32]) {
-        self.input.extend(input);
+        self.prover.add_input(input)
     }
 
     /// Add a digest to the input for the recursion program.
     fn add_input_digest(&mut self, digest: &Digest, kind: DigestKind) {
-        match kind {
-            // Poseidon2 digests consist of  BabyBear field elems and do not need to be split.
-            DigestKind::Poseidon2 => self.add_input(digest.as_words()),
-            // SHA-256 digests need to be split into 16-bit half words to avoid overflowing.
-            DigestKind::Sha256 => self.add_input(bytemuck::cast_slice(
-                &digest
-                    .as_words()
-                    .iter()
-                    .copied()
-                    .flat_map(|x| [x & 0xffff, x >> 16])
-                    .map(BabyBearElem::new)
-                    .collect::<Vec<_>>(),
-            )),
-        }
+        self.prover.add_input_digest(digest, kind)
     }
 
     /// Add a recursion seal (i.e. STARK proof) to input tape of the recursion program.
@@ -734,25 +559,7 @@ impl Prover {
     /// program and input.
     #[tracing::instrument(skip_all)]
     pub fn run(&mut self) -> Result<RecursionReceipt> {
-        // NOTE: Code is repeated across match arms to satisfy generics.
-        match self.opts.hashfn.as_ref() {
-            "poseidon2" => {
-                let hal_pair = poseidon2_hal_pair();
-                let (hal, circuit_hal) = (hal_pair.hal.as_ref(), hal_pair.circuit_hal.as_ref());
-                self.run_with_hal(hal, circuit_hal)
-            }
-            "poseidon_254" => {
-                let hal_pair = poseidon254_hal_pair();
-                let (hal, circuit_hal) = (hal_pair.hal.as_ref(), hal_pair.circuit_hal.as_ref());
-                self.run_with_hal(hal, circuit_hal)
-            }
-            "sha-256" => {
-                let hal_pair = sha256_hal_pair();
-                let (hal, circuit_hal) = (hal_pair.hal.as_ref(), hal_pair.circuit_hal.as_ref());
-                self.run_with_hal(hal, circuit_hal)
-            }
-            _ => bail!("no hal found for {}", self.opts.hashfn),
-        }
+        self.prover.run()
     }
 
     /// Run the prover, producing a receipt of execution for the recursion circuit over the loaded
@@ -763,99 +570,6 @@ impl Prover {
         H: Hal<Field = BabyBear, Elem = BabyBearElem, ExtElem = BabyBearExtElem>,
         C: CircuitHal<H>,
     {
-        let machine_ctx = self.preflight()?;
-
-        let split_points = core::mem::take(&mut self.split_points);
-
-        let mut executor =
-            exec::RecursionExecutor::new(&CIRCUIT, &self.program, machine_ctx, split_points);
-        executor.run()?;
-
-        let mut adapter = ProveAdapter::new(&mut executor.executor);
-        let mut prover = risc0_zkp::prove::Prover::new(hal, CIRCUIT.get_taps());
-        let hashfn = Rc::clone(&hal.get_hash_suite().hashfn);
-
-        // At the start of the protocol, seed the Fiat-Shamir transcript with context information
-        // about the proof system and circuit.
-        prover
-            .iop()
-            .commit(&hashfn.hash_elem_slice(&PROOF_SYSTEM_INFO.encode()));
-        prover
-            .iop()
-            .commit(&hashfn.hash_elem_slice(&CircuitImpl::CIRCUIT_INFO.encode()));
-
-        adapter.execute(prover.iop(), hal);
-
-        prover.set_po2(adapter.po2() as usize);
-
-        let ctrl = hal.copy_from_elem("ctrl", &adapter.get_code().as_slice());
-        prover.commit_group(REGISTER_GROUP_CTRL, &ctrl);
-
-        let data = hal.copy_from_elem("data", &adapter.get_data().as_slice());
-        prover.commit_group(REGISTER_GROUP_DATA, &data);
-
-        // Make the mixing values
-        let mix: Vec<_> = (0..CircuitImpl::MIX_SIZE)
-            .map(|_| prover.iop().random_elem())
-            .collect();
-        let mix = hal.copy_from_elem("mix", mix.as_slice());
-
-        let steps = adapter.get_steps();
-        let mut accum = vec![BabyBearElem::INVALID; steps * CIRCUIT.accum_size()];
-
-        // Add random noise to end of accum
-        let mut rng = thread_rng();
-        for i in steps - ZK_CYCLES..steps {
-            for j in 0..CIRCUIT.accum_size() {
-                accum[j * steps + i] = BabyBearElem::random(&mut rng);
-            }
-        }
-
-        let io = hal.copy_from_elem("io", &adapter.get_io().as_slice());
-        let accum = hal.copy_from_elem("accum", accum.as_slice());
-
-        circuit_hal.accumulate(&ctrl, &io, &data, &mix, &accum, steps);
-
-        prover.commit_group(REGISTER_GROUP_ACCUM, &accum);
-
-        let seal = prover.finalize(&[&mix, &io], circuit_hal);
-
-        Ok(RecursionReceipt {
-            control_id: self.control_id,
-            seal,
-            output: self.output.clone(),
-        })
-    }
-
-    #[tracing::instrument(skip_all)]
-    fn preflight(&mut self) -> Result<exec::MachineContext> {
-        let mut machine = exec::MachineContext::new(take(&mut self.input));
-        let mut preflight = preflight::Preflight::new(&mut machine);
-
-        for (cycle, row) in self.program.code_by_row().enumerate() {
-            let ctx = CircuitStepContext {
-                cycle,
-                size: (1 << RECURSION_PO2) - ZK_CYCLES,
-            };
-
-            preflight.set_top(&ctx, row)?
-        }
-
-        // TODO: is this necessary?
-        let zero_row = vec![BabyBearElem::ZERO; self.program.code_size];
-        for cycle in self.program.code_rows()..(1 << RECURSION_PO2) - ZK_CYCLES {
-            let ctx = CircuitStepContext {
-                cycle,
-                size: (1 << RECURSION_PO2) - ZK_CYCLES,
-            };
-
-            preflight.set_top(&ctx, &zero_row)?
-        }
-
-        self.split_points = preflight.split_points;
-        self.split_points.push((1 << RECURSION_PO2) - ZK_CYCLES);
-        self.output = preflight.output;
-        machine.iop_reads = preflight.iop_reads;
-        Ok(machine)
+        self.prover.run_with_hal(hal, circuit_hal)
     }
 }
