+++ risc0/zkvm/src/receipt/succinct.rs
@@ -26,7 +26,7 @@ use risc0_zkp::{
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    receipt::{merkle::MerkleProof, VerifierContext},
+    receipt::{MerkleProof, VerifierContext},
     receipt_claim::{MaybePruned, Unknown},
     sha,
 };
